Use iterators for prefix search to reduce peak memory usage. Code with
iterators has slightly worse performance for search requests that have
small number of matches. However, iterators significantly reduce memory
usage for cases when there are a lot of matches or we have to scan
all entries.

Also replace `uint64`/`float64` with `uint32`/`float32` - it improves
performance and reduces memory usage even further.

```
                          │ slices.txt  │             iter.txt              │          iter-uint32.txt          │
                          │   sec/op    │   sec/op     vs base              │   sec/op     vs base              │
vendor-32                   262.8µ ± 1%   268.0µ ± 1%  +1.99% (p=0.001 n=7)   261.5µ ± 1%       ~ (p=0.128 n=7)
vendor_-github-32           228.1µ ± 1%   234.2µ ± 1%  +2.71% (p=0.001 n=7)   228.8µ ± 1%       ~ (p=0.165 n=7)
"vendor"_-github-32         281.2µ ± 1%   278.9µ ± 1%  -0.83% (p=0.038 n=7)   282.9µ ± 3%       ~ (p=0.456 n=7)
"vendor"_github_google-32   599.8µ ± 2%   609.3µ ± 2%       ~ (p=0.128 n=7)   570.4µ ± 2%  -4.91% (p=0.001 n=7)
rview-32                    2.729µ ± 2%   2.946µ ± 2%  +7.95% (p=0.001 n=7)   2.889µ ± 1%  +5.86% (p=0.001 n=7)
Dockerfile-32               19.66µ ± 1%   19.62µ ± 1%       ~ (p=0.620 n=7)   19.05µ ± 1%  -3.10% (p=0.001 n=7)
"Dockerfile"-32             196.3µ ± 2%   206.4µ ± 1%  +5.19% (p=0.001 n=7)   206.9µ ± 2%  +5.42% (p=0.001 n=7)
geomean                     100.9µ        103.5µ       +2.59%                 101.4µ       +0.46%

                          │  slices.txt   │              iter.txt               │           iter-uint32.txt           │
                          │     B/op      │     B/op      vs base               │     B/op      vs base               │
vendor-32                    176.2Ki ± 0%   229.4Ki ± 0%  +30.17% (p=0.001 n=7)   193.6Ki ± 0%   +9.87% (p=0.001 n=7)
vendor_-github-32            164.2Ki ± 0%   173.6Ki ± 0%   +5.69% (p=0.001 n=7)   137.8Ki ± 0%  -16.10% (p=0.001 n=7)
"vendor"_-github-32         228.08Ki ± 0%   63.81Ki ± 0%  -72.02% (p=0.001 n=7)   63.81Ki ± 0%  -72.02% (p=0.001 n=7)
"vendor"_github_google-32    876.6Ki ± 0%   877.2Ki ± 0%   +0.07% (p=0.001 n=7)   591.8Ki ± 0%  -32.49% (p=0.001 n=7)
rview-32                     1.797Ki ± 0%   2.344Ki ± 0%  +30.43% (p=0.001 n=7)   2.141Ki ± 0%  +19.13% (p=0.001 n=7)
Dockerfile-32                12.52Ki ± 0%   13.81Ki ± 0%  +10.36% (p=0.001 n=7)   11.78Ki ± 0%   -5.87% (p=0.001 n=7)
"Dockerfile"-32             205120.0 ± 0%     752.0 ± 0%  -99.63% (p=0.001 n=7)     752.0 ± 0%  -99.63% (p=0.001 n=7)
geomean                      82.52Ki        34.04Ki       -58.75%                 29.33Ki       -64.46%

                          │ slices.txt │              iter.txt               │           iter-uint32.txt           │
                          │ allocs/op  │  allocs/op   vs base                │  allocs/op   vs base                │
vendor-32                   59.00 ± 0%    76.00 ± 0%   +28.81% (p=0.001 n=7)    76.00 ± 0%   +28.81% (p=0.001 n=7)
vendor_-github-32           60.00 ± 0%    79.00 ± 0%   +31.67% (p=0.001 n=7)    79.00 ± 0%   +31.67% (p=0.001 n=7)
"vendor"_-github-32         5.000 ± 0%   22.000 ± 0%  +340.00% (p=0.001 n=7)   22.000 ± 0%  +340.00% (p=0.001 n=7)
"vendor"_github_google-32   172.0 ± 0%    182.0 ± 0%    +5.81% (p=0.001 n=7)    182.0 ± 0%    +5.81% (p=0.001 n=7)
rview-32                    22.00 ± 0%    32.00 ± 0%   +45.45% (p=0.001 n=7)    32.00 ± 0%   +45.45% (p=0.001 n=7)
Dockerfile-32               57.00 ± 0%    69.00 ± 0%   +21.05% (p=0.001 n=7)    69.00 ± 0%   +21.05% (p=0.001 n=7)
"Dockerfile"-32             6.000 ± 0%   13.000 ± 0%  +116.67% (p=0.001 n=7)   13.000 ± 0%  +116.67% (p=0.001 n=7)
geomean                     30.20         49.12        +62.66%                  49.12        +62.66%
```

+++ web/models.go
@@ -71,7 +71,7 @@ type SearchHit struct {
 	IsDir   bool    `json:"is_dir"`
 	Size    int64   `json:"size"`
 	ModTime int64   `json:"mod_time"`
-	Score   float64 `json:"score"`
+	Score   float32 `json:"score"`
 	WebURL  string  `json:"web_url"`
 	Icon    string  `json:"icon"`
 }
